refactor: grouped other common methods between modules
