+++ src/qibolab/instruments/qblox/cluster_qcm_bb.py
@@ -254,20 +254,6 @@ class ClusterQCM_BB(ClusterModule):
         sequencer.qubit = qubit.name if qubit else None
         return sequencer
 
-    def get_if(self, pulse):
-        """Returns the intermediate frequency needed to synthesise a pulse
-        based on the port lo frequency."""
-
-        _rf = pulse.frequency
-        _lo = 0  # QCMs do not have local oscillator
-        _if = _rf - _lo
-        if abs(_if) > self.FREQUENCY_LIMIT:
-            raise RuntimeError(
-                f"""
-            Pulse frequency {_rf:_} cannot be synthesised, it exceeds the maximum frequency of {self.FREQUENCY_LIMIT:_}"""
-            )
-        return _if
-
     def process_pulse_sequence(
         self,
         qubits: dict,
@@ -735,17 +721,6 @@ class ClusterQCM_BB(ClusterModule):
             with open(filename, "w", encoding="utf-8") as file:
                 print_readable_snapshot(self.device, file, update=True)
 
-    def play_sequence(self):
-        """Executes the sequence of instructions."""
-
-        for sequencer_number in self._used_sequencers_numbers:
-            # Start used sequencers
-            self.device.start_sequencer(sequencer_number)
-
-    def start(self):
-        """Empty method to comply with Instrument interface."""
-        pass
-
     def stop(self):
         """Stops all sequencers."""
 
@@ -766,8 +741,3 @@ class ClusterQCM_BB(ClusterModule):
 
         except:
             log.warning("Unable to clear offsets")
-
-    def disconnect(self):
-        """Empty method to comply with Instrument interface."""
-        self.is_connected = False
-        self.device = None
